Refactor UserContext with useReducer
